+++ src/components/MoviesCard/MoviesCard.js
@@ -1,52 +1,10 @@
 import React, { useState } from "react";
-import { useLocation } from "react-router-dom";
 import "./MoviesCard.css";
-import { mainApi } from "../../utils/MainApi";
 
 const MoviesCard = ({ movie }) => {
   const [checked, setChecked] = useState(false);
   const { nameRU, duration, image, trailerLink } = movie;
   const { url } = image;
-
-  const location = useLocation();
-
-  const handleSaveMovie = (movie) => {
-    mainApi
-      .createMovie(movie)
-      .then((res) => console.log(res))
-      .catch((e) => console.log(e))
-      .finally();
-  };
-
-  const handleDeleteMovie = (id) => {
-    mainApi
-      .deleteMovie(id)
-      .then((res) => console.log(res))
-      .catch((e) => console.log(e))
-      .finally();
-  };
-
-  const handleFavoriteClick = (movie) => {
-    const newMovie = {
-      country: movie.country || "Неизвестно",
-      director: movie.director,
-      duration: movie.duration,
-      year: movie.year,
-      description: movie.description,
-      trailerLink: movie.trailerLink,
-      nameRU: movie.nameRU,
-      nameEN: movie.nameEN,
-      image: `https://api.nomoreparties.co${url}`,
-      thumbnail: `https://api.nomoreparties.co${image.formats.thumbnail.url}`,
-      movieId: movie.id,
-    };
-    console.log(movie);
-    if (location.pathname === "/movies") {
-      handleSaveMovie(newMovie);
-    } else {
-      handleDeleteMovie(movie.id);
-    }
-  };
 
   const durationSorted = () => {
     const m = duration % 60;
@@ -67,7 +25,6 @@ const MoviesCard = ({ movie }) => {
             type="checkbox"
             checked={checked}
             onChange={() => setChecked((prev) => !prev)}
-            onClick={() => handleFavoriteClick(movie)}
           />
         </div>
       </div>
